requirements.txt
torch==2.0.0                      
torchvision==0.15.0               
transformers==4.29.0             
pillow==9.5.0                    
opencv-python==4.7.0             
numpy==1.23.5                    
pandas==2.0.3                    
scikit-learn==1.2.2               
tesseract==0.3.0                  
ctpn==0.1.0                       
matplotlib==3.7.1                 
seaborn==0.12.2                   
